+++ backend/MainServer/src/main/java/com/a501/recipe/api/controller/IntakeController.java
@@ -1,6 +1,7 @@
 package com.a501.recipe.api.controller;
 
 import com.a501.recipe.aop.LoginUser;
+import com.a501.recipe.aop.exception.AccessDeniedException;
 import com.a501.recipe.api.domain.entity.User;
 import com.a501.recipe.api.dto.intake.IntakeAddRequestDto;
 import com.a501.recipe.api.dto.intake.IntakeDto;
@@ -9,6 +10,7 @@ import com.a501.recipe.api.dto.response.CommonResult;
 import com.a501.recipe.api.dto.response.ManyResult;
 import com.a501.recipe.api.service.IntakeService;
 import com.a501.recipe.api.service.ResponseService;
+import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -17,9 +19,10 @@ import springfox.documentation.annotations.ApiIgnore;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
+@Api(tags = "07. Intake Controller")
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/intake")
+@RequestMapping("/user/{userId}/intake")
 public class IntakeController {
 
     private final IntakeService intakeService;
@@ -27,7 +30,10 @@ public class IntakeController {
 
     @ApiOperation(value = "특정 날짜 이상부터 섭취한 음식 조회")
     @GetMapping()
-    public ManyResult<IntakeDto> getMyIntakeList(@ApiIgnore @LoginUser User loginUser, @RequestParam("date") String date) {
+    public ManyResult<IntakeDto> getMyIntakeList(@ApiIgnore @LoginUser User loginUser,
+                                                 @PathVariable("userId") Long userId,
+                                                 @RequestParam("date") String date) {
+        if (!userId.equals(loginUser.getId())) throw new AccessDeniedException();
         LocalDate baseDate = LocalDate.parse(date);
         return responseService.getManyResult(intakeService.getMyIntakeList(loginUser, baseDate));
     }
@@ -35,7 +41,9 @@ public class IntakeController {
     @ApiOperation(value = "섭취한 음식 추가")
     @PostMapping()
     public CommonResult addIntake(@ApiIgnore @LoginUser User loginUser,
+                                  @PathVariable("userId") Long userId,
                                   @RequestBody IntakeAddRequestDto intakeAddRequestDto) {
+        if (!userId.equals(loginUser.getId())) throw new AccessDeniedException();
         intakeService.addIntake(loginUser, intakeAddRequestDto);
         return responseService.getSuccessResult();
     }
@@ -43,8 +51,10 @@ public class IntakeController {
     @ApiOperation(value = "섭취한 음식 수정")
     @PutMapping("/{id}")
     public CommonResult updateIntake(@ApiIgnore @LoginUser User loginUser,
-                                  @PathVariable("id") Long id,
-                                  @RequestBody IntakeUpdateRequestDto intakeUpdateRequestDto) throws IllegalAccessException {
+                                     @PathVariable("userId") Long userId,
+                                     @PathVariable("id") Long id,
+                                     @RequestBody IntakeUpdateRequestDto intakeUpdateRequestDto)  {
+        if (!userId.equals(loginUser.getId())) throw new AccessDeniedException();
         intakeService.updateIntake(loginUser, id, intakeUpdateRequestDto);
         return responseService.getSuccessResult();
     }
@@ -52,11 +62,12 @@ public class IntakeController {
     @ApiOperation(value = "섭취한 음식 삭제")
     @DeleteMapping("/{id}")
     public CommonResult updateIntake(@ApiIgnore @LoginUser User loginUser,
-                                     @PathVariable("id") Long id) throws IllegalAccessException {
+                                     @PathVariable("userId") Long userId,
+                                     @PathVariable("id") Long id) {
+        if (!userId.equals(loginUser.getId())) throw new AccessDeniedException();
         intakeService.deleteIntake(loginUser, id);
         return responseService.getSuccessResult();
     }
-
 
 
 }
